+++ src/main/java/de/maxhenkel/advancedtools/MixinConnector.java
@@ -1,0 +1,14 @@
+package de.maxhenkel.advancedtools;
+
+import org.spongepowered.asm.mixin.Mixins;
+import org.spongepowered.asm.mixin.connect.IMixinConnector;
+
+public class MixinConnector implements IMixinConnector {
+
+    @Override
+    public void connect() {
+        Mixins.addConfiguration("assets/" + Main.MODID + "/advancedtools.mixins.json");
+    }
+
+}
+